+++ src/dynamics/time_integration.jl
@@ -291,7 +291,7 @@ function timestep!(
 
     # PARTICLE ADVECTION (always skip 1st step of first_timesteps!)
     not_first_timestep = lf2 == 2
-    not_first_timestep && particle_advection!(progn, diagn, model.particle_advection)
+    not_first_timestep && particle_advection!(progn, diagn, model)
 
     return nothing 
 end
@@ -323,7 +323,7 @@ function timestep!(
     
     # PARTICLE ADVECTION (always skip 1st step of first_timesteps!)
     not_first_timestep = lf2 == 2
-    not_first_timestep && particle_advection!(progn, diagn, model.particle_advection)
+    not_first_timestep && particle_advection!(progn, diagn, model)
 
     return nothing
 end
@@ -370,7 +370,7 @@ function timestep!(
 
     # PARTICLE ADVECTION (always skip 1st step of first_timesteps!)
     not_first_timestep = lf2 == 2
-    not_first_timestep && particle_advection!(progn, diagn, model.particle_advection)
+    not_first_timestep && particle_advection!(progn, diagn, model)
 
     return nothing 
 end